+++ bundles/org.openhab.binding.magentatv/pom.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-	<modelVersion>4.0.0</modelVersion>
+  <modelVersion>4.0.0</modelVersion>
 
-	<parent>
-		<groupId>org.openhab.addons.bundles</groupId>
-		<artifactId>org.openhab.addons.reactor.bundles</artifactId>
-		<version>2.5.9-SNAPSHOT</version>
-	</parent>
+  <parent>
+    <groupId>org.openhab.addons.bundles</groupId>
+    <artifactId>org.openhab.addons.reactor.bundles</artifactId>
+    <version>2.5.9-SNAPSHOT</version>
+  </parent>
 
-	<artifactId>org.openhab.binding.magentatv</artifactId>
+  <artifactId>org.openhab.binding.magentatv</artifactId>
 
-	<name>openHAB Add-ons :: Bundles :: MagentaTV Binding</name>
+  <name>openHAB Add-ons :: Bundles :: MagentaTV Binding</name>
 
 </project>
