+++ bundles/org.openhab.binding.magentatv/src/main/java/org/openhab/binding/magentatv/internal/MagentaTVHandlerFactory.java
@@ -304,5 +304,4 @@ public class MagentaTVHandlerFactory extends BaseThingHandlerFactory {
             logger.debug("Unable to process SSDP message for IP {} - {}", ipAddress, e.toString());
         }
     }
-
 }
