+++ bundles/org.openhab.binding.magentatv/src/main/java/org/openhab/binding/magentatv/internal/network/MagentaTVPoweroffListener.java
@@ -106,7 +106,6 @@ public class MagentaTVPoweroffListener extends Thread {
         } finally {
             close();
         }
-
     }
 
     public boolean isStarted() {
