+++ bundles/org.openhab.binding.magentatv/src/main/resources/ESH-INF/thing/thing-types.xml
@@ -22,32 +22,11 @@
 
 
 		<config-description>
-			<parameter name="udn" type="text">
-				<label>Unique Device Name</label>
-				<description>The UDN identifies the Media Receiver</description>
-				<required>true</required>
-			</parameter>
-			<parameter name="modelId" type="text">
-				<label>Receiver Model</label>
-				<description>Model type (MR4xx/ MR2xx)</description>
-				<required>true</required>
-				<options>
-					<option value="Auto">AUTO</option>
-					<option value="DMS_TPB">MR400</option>
-					<option value="MR401B">MR401B</option>
-					<option value="MR201">MR201</option>
-				</options>
-			</parameter>
 			<parameter name="ipAddress" type="text">
 				<label>Media Receiver's IP Address</label>
 				<description>IP address of the receiver</description>
 				<required>true</required>
 				<context>network-address</context>
-			</parameter>
-			<parameter name="port" type="text">
-				<label>Receiver Port</label>
-				<description>Port address for UPnP</description>
-				<required>true</required>
 			</parameter>
 			<parameter name="accountName" type="text">
 				<label>Account Name</label>
@@ -59,6 +38,18 @@
 				<description>Credentials: Account Password (same as for the Telekom Kundencenter)</description>
 				<context>password</context>
 				<required>true</required>
+			</parameter>
+			<parameter name="udn" type="text">
+				<label>Unique Device Name</label>
+				<description>The UDN identifies the Media Receiver</description>
+				<required>true</required>
+				<advanced>true</advanced>
+			</parameter>
+			<parameter name="port" type="text">
+				<label>Receiver Port</label>
+				<description>Port address for UPnP</description>
+				<required>true</required>
+				<advanced>true</advanced>
 			</parameter>
 			<parameter name="userId" type="text">
 				<label>UID</label>
